Added a section using Ridge Regression and another on inflation rates
Added a section that uses Ridge Regression to try and reduce the
variance in the polynomial model to make it a bit more reliable. I
also added a quick look at the inflation rates over the past 27 years
to see how they correspond to increases in the tuition at Stanford.