Updated my version of ipynb plugin
I've udpated my local version of the ipynb plugin for pelican so that
it uses the typical nbconvert config found in the user's home
directory, or in the current project directory, etc. I've got a pull
request in for this change, so hopefully, I can simply update to the
latest version of the plugin soon and drop this PR branch.